volunteer_app_backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
django==4.2.13
    # via
    #   -r requirements.in
    #   django-filter
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-filter==24.2
    # via -r requirements.in
django-phonenumber-field==7.3.0
    # via -r requirements.in
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-spectacular==0.27.2
    # via -r requirements.in
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.22.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
phonenumbers==8.13.36
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via -r requirements.in
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.0
    # via django
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.1
    # via requests
